chore: more documentation
